--- external_modules/shared-context/walcron-zelda-shared-context.js
@@ -1,2 +1,0 @@
-System.register(["rxjs"],(function(e,r){var o={};return{setters:[function(e){o.BehaviorSubject=e.BehaviorSubject}],execute:function(){e((()=>{var e={722:(e,r,o)=>{const t=o(905).R;r.s=function(e){if(e||(e=1),!o.y.meta||!o.y.meta.url)throw console.error("__system_context__",o.y),Error("systemjs-webpack-interop was provided an unknown SystemJS context. Expected context.meta.url, but none was provided");o.p=t(o.y.meta.url,e)}},905:(e,r,o)=>{r.R=function(e,r){var o=document.createElement("a");o.href=e;for(var t="/"===o.pathname[0]?o.pathname:"/"+o.pathname,n=0,s=t.length;n!==r&&s>=0;)"/"===t[--s]&&n++;if(n!==r)throw Error("systemjs-webpack-interop: rootDirectoryLevel ("+r+") is greater than the number of directories ("+n+") in the URL path "+e);var i=t.slice(0,s+1);return o.protocol+"//"+o.host+i};Number.isInteger},792:e=>{"use strict";e.exports=o}},t={};function n(r){var o=t[r];if(void 0!==o)return o.exports;var s=t[r]={exports:{}};return e[r](s,s.exports,n),s.exports}n.y=r,n.d=(e,r)=>{for(var o in r)n.o(r,o)&&!n.o(e,o)&&Object.defineProperty(e,o,{enumerable:!0,get:r[o]})},n.o=(e,r)=>Object.prototype.hasOwnProperty.call(e,r),n.r=e=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.p="";var s={};return(0,n(722).s)(1),(()=>{"use strict";n.r(s),n.d(s,{INTERVAL_CHECK_IN_MILISECONDS:()=>r,auth$:()=>o,login:()=>i,logout:()=>a,name:()=>u});var e=n(792),r=2500,o=new e.BehaviorSubject({sessionToken:localStorage.getItem("sessionToken"),error:void 0,pending:!1}),t=function(e,t){return new Promise((function(n,s){o.next({sessionToken:null,error:void 0,pending:!0}),setTimeout((function(){if("exampleuser"===e&&"examplepassword"===t){var r="abc123def456";localStorage.setItem("sessionToken",r),n({sessionToken:r,error:void 0,pending:!1})}else n({sessionToken:null,error:"Invalid user or password",pending:!1})}),r)}))};function i(e,r){o.value.pending||t(e,r).then((function(e){o.next(e)}))}function a(){localStorage.removeItem("sessionToken"),o.next({sessionToken:null,error:void 0,pending:!1})}var u="shared-context"})(),s})())}}}));
-//# sourceMappingURL=walcron-zelda-shared-context.js.map
